Feature/vosa 240 repackaging live center and other plugins that require patching of webservice.war (and/or other ECE WARs). Every WAR, except escenic-admin.war and escenic.war can now be patched by WARs in the EAR file.

+++ usr/share/escenic/ece-scripts/ece.d/repackage.sh
@@ -39,24 +39,18 @@ get_installed_packages_studio_list() {
 }
 
 ## $1 :: tmp dir with EAR contents
-remove_webapps_that_are_included_in_installed_packages() {
+copy_webapps_included_in_installed_packages_if_not_exists() {
   local tmp_dir=$1
-  for d in $(get_installed_packages_war_list); do
-    for package_war in $(find "${d}" -type f -name "*.war"); do
-      for ear_war in "${tmp_dir}/"*.war; do
-        if [[ $(basename "${ear_war}") == $(basename "${package_war}") ]]; then
-          run rm "${ear_war}"
-        fi
-      done
-    done
-  done
-}
-
-## $1 :: tmp dir with EAR contents
-copy_webapps_that_are_included_in_installed_packages() {
-  local tmp_dir=$1
+  local war=
   for war in $(get_installed_packages_war_list); do
-    run cp "${war}" "${tmp_dir}"
+    local war_base=
+    war_base=$(basename "${war}")
+    if [[ ! -e "${tmp_dir}/${war_base}" ]]; then
+      run cp "${war}" "${tmp_dir}"
+    else
+      debug "EAR already has ${war_base}, will patch it"
+      # will be patched in overwrite_webapp_libs(), see main()
+    fi
   done
 }
 
@@ -69,6 +63,8 @@ get_file_base() {
   basename "$1" "$2" |
     sed -r \
         -e 's#[-][0-9.-]+##' \
+        -e 's#-trunk-SNAPSHOT##' \
+        -e 's#-develop-SNAPSHOT##' \
         -e 's#SNAPSHOT##' \
         -e 's#[.]$##' \
         -e 's#(alpha|beta|rc)[.-][0-9]+##'
@@ -77,9 +73,12 @@ get_file_base() {
 ## $1 :: tmp dir with EAR contents
 remove_libs_included_in_installed_packages() {
   local tmp_dir=$1
+
+  local package_jar=
   for package_jar in $(get_installed_packages_lib_list); do
     package_jar_base=$(get_file_base "${package_jar}" jar)
 
+    local ear_jar=
     for ear_jar in "${tmp_dir}/lib/"*.jar; do
       ear_jar_base=$(get_file_base "${ear_jar}" jar)
       if [[ "${ear_jar_base}" == "${package_jar_base}" ]]; then
@@ -96,6 +95,7 @@ copy_libs_included_in_installed_packages() {
   local tmp_dir=$1
   make_dir "${tmp_dir}/lib"
 
+  local package_jar=
   for package_jar in $(get_installed_packages_lib_list); do
     debug "Including JAR ${package_jar} in EAR ..."
     run cp "${package_jar}" "${tmp_dir}/lib"
@@ -103,37 +103,45 @@ copy_libs_included_in_installed_packages() {
 }
 
 ## $1 :: tmp dir with WAR contents
-overwrite_publication_webapp_libs_included_in_installed_packages() {
+overwrite_webapp_libs_included_in_installed_packages() {
   local tmp_dir=$1
 
-  for pub_war_jar in "${tmp_dir}/WEB-INF/lib/"*.jar; do
-    pub_war_jar_base=$(get_file_base "${pub_war_jar}" jar)
-    debug "${FUNCNAME[0]}
-      Checking pub_war_jar=$pub_war_jar
-      pub_war_jar_base=$pub_war_jar_base"
-
-    for package_dir in "${USR_SHARE_DIR}/"*; do
-      local template_lib_dir="${package_dir}/template/WEB-INF/lib"
+  local package_dir=
+  for package_dir in "${USR_SHARE_DIR}/"*; do
+    local template_lib_dir="${package_dir}/template/WEB-INF/lib"
+    if [ ! -d "${template_lib_dir}" ]; then
+      # ECE and its plugins put their publication specific libraries
+      # in different directory structures.
+      template_lib_dir="${package_dir}/publication/webapp/WEB-INF/lib"
       if [ ! -d "${template_lib_dir}" ]; then
-        # ECE and its plugins put their publication specific libraries
-        # in different directory structures.
-        template_lib_dir="${package_dir}/publication/webapp/WEB-INF/lib"
-        if [ ! -d "${template_lib_dir}" ]; then
-          continue
-        fi
+        continue
       fi
+    fi
 
-      find "${template_lib_dir}" -name "*.jar" -type f | \
-        while read -r package_jar; do
-          package_jar_base=$(get_file_base "${package_jar}" jar)
+    find "${template_lib_dir}" -name "*.jar" -type f | \
+      while read -r package_jar; do
+        package_jar_base=$(get_file_base "${package_jar}" jar)
+
+        local pub_war_jar=
+        for pub_war_jar in "${tmp_dir}/WEB-INF/lib/"*.jar; do
+          pub_war_jar_base=$(get_file_base "${pub_war_jar}" jar)
+          debug "${FUNCNAME[0]}
+                  Checking pub_war_jar=$pub_war_jar
+                  pub_war_jar_base=$pub_war_jar_base"
 
           if [[ "${pub_war_jar_base}" == "${package_jar_base}" ]]; then
             debug " -> Replacing ${pub_war_jar} with ${package_jar}"
             run rm "${pub_war_jar}"
             run cp "${package_jar}" "$(dirname "${pub_war_jar}")"
+          else
+            debug " -> ${pub_war} doesn't have ${package_jar}, adding it"
+            local dir=
+            dir="$(dirname "${pub_war_jar}")"
+            make_dir "${dir}"
+            run cp "${package_jar}" "${dir}/."
           fi
         done
-    done
+      done
   done
 }
 
@@ -161,39 +169,58 @@ create_new_ear_in_dir() {
   echo "${new_ear}"
 }
 
-## Overwrite the JAR libraries inside the EAR specific webapps,
-## most/all of which will be publication webapps.
+## $1 :: war
+should_patch_war() {
+  local war=$1
+  local war_base=
+  war_base=$(basename "${war}")
+
+  local -a should_not_be_patched=(
+    "escenic-admin.war"
+    "escenic.war"
+  )
+
+  local el=
+  for el in "${should_not_be_patched[@]}"; do
+    if [[ "${el}" == "${war_base}" ]]; then
+      return 1
+    fi
+  done
+
+  return 0
+}
+
+## Overwrite the JAR libraries inside the EAR specific webapps.
 ##
 ## $1 :: the dir in which the extracted EAR resides
-overwrite_publication_webapp_libs() {
+overwrite_webapp_libs() {
   local dir=$1
   local installed_package_war_list=
   installed_package_war_list=$(get_installed_packages_war_list)
 
-  for war in ${dir}/*.war; do
-    if [[ $(echo "${installed_package_war_list}" | grep -c "$(basename "${war}")") -gt 0 ]]; then
-      log "Don't need to fix $(basename "${war}") (it's provided by a package)"
-    else
-      local start=
-      local tmp_dir=
-      tmp_dir=$(mktemp -d)
-
-      log "Fixing $(basename "${war}"), extracting it to ${tmp_dir} ..."
-      run unzip -q "${war}" -d "${tmp_dir}"
-
-      start=$(date +%s)
-      overwrite_publication_webapp_libs_included_in_installed_packages "${tmp_dir}"
-      log "Fixing $(basename "${war}") took $(( $(date +%s) - start)) seconds ..."
-
-      # to be sure it's prestine, we delete the old war
-      run rm "${war}"
-      (
-        cd "${tmp_dir}" || exit 1
-        run zip --recurse-paths --quiet "${war}" .
-      )
-
-      run rm -rf "${tmp_dir}"
+  find "${dir}" -name "*.war" | while read -r war; do
+    if ! should_patch_war "${war}"; then
+      continue
     fi
+    local start=
+    local tmp_dir=
+    tmp_dir=$(mktemp -d)
+
+    log "Fixing $(basename "${war}"), extracting it to ${tmp_dir} ..."
+    run unzip -q "${war}" -d "${tmp_dir}"
+
+    start=$(date +%s)
+    overwrite_webapp_libs_included_in_installed_packages "${tmp_dir}"
+    log "Fixing $(basename "${war}") took $(( $(date +%s) - start)) seconds ..."
+
+    # to be sure it's prestine, we delete the old war
+    run rm "${war}"
+    (
+      cd "${tmp_dir}" || exit 1
+      run zip --recurse-paths --quiet "${war}" .
+    )
+
+    run rm -rf "${tmp_dir}"
   done
 }
 
@@ -304,6 +331,7 @@ merge_all_dirs_to_war() {
     remove_pid_and_exit_in_error
   fi
 
+  local dir=
   for dir in ${dir_list}; do
     if [ ! -d "${dir}/webapp" ]; then
       continue
@@ -336,6 +364,7 @@ merge_all_plugins_with_studio_war() {
     remove_pid_and_exit_in_error
   fi
 
+  local dir=
   for dir in ${dir_list}; do
     log "Adding studio plugin in ${dir} ..."
     local plugin_name=
@@ -417,13 +446,11 @@ repackage() {
   start_time=$(date +%s)
   extract_archive "${ear}" "${tmp_dir}"
 
-  remove_webapps_that_are_included_in_installed_packages "${tmp_dir}"
-  copy_webapps_that_are_included_in_installed_packages "${tmp_dir}"
-
   remove_libs_included_in_installed_packages "${tmp_dir}"
   copy_libs_included_in_installed_packages "${tmp_dir}"
 
-  overwrite_publication_webapp_libs "${tmp_dir}"
+  copy_webapps_included_in_installed_packages_if_not_exists "${tmp_dir}"
+  overwrite_webapp_libs "${tmp_dir}"
   merge_all_webservice_extension_webapps "${tmp_dir}"
   merge_all_webservice_webapps "${tmp_dir}"
   merge_all_studio_plugins "${tmp_dir}"
